% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.Bolstad.R
\name{as.data.frame.Bolstad}
\alias{as.data.frame.Bolstad}
\title{as.data.frame.Bolstad}
\usage{
\method{as.data.frame}{Bolstad}(x, ...)
}
\arguments{
\item{x}{an object of class \code{Bolstad}}

\item{\dots, }{any extra arguments needed.}
}
\description{
as.data.frame.Bolstad
}

