% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayes.lm.R
\name{bayes.lm}
\alias{bayes.lm}
\title{Bayesian inference for multiple linear regression}
\usage{
bayes.lm(formula, data, subset, na.action, model = TRUE, x = FALSE,
  y = FALSE, center = TRUE, prior = NULL)
}
\arguments{
\item{formula}{an object of class \code{\link[stats]{formula}} (or one that can be coerced to that class): a symbolic
description of the model to be fitted. The details of model specification are given under `Details'.}

\item{data}{an optional data frame, list or environment (or object coercible by \code{\link[stats]{as.data.frame}} to a
data frame) containing the variables in the model. If not found in data, the variables are taken
from \code{environment(formula)}, typically the environment from which \code{bayes.lm} is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{na.action}{a function which indicates what should happen when the data contain \code{NA}s. The
default is set by the \code{\link[stats]{na.action}} setting of options, and is \code{link[stats]{na.fail}}
 if that is unset. The `factory-fresh' default is \code{\link[stats]{na.omit}}. Another possible value
 is \code{NULL}, no action. Value \code{\link[stats]{na.exclude}} can be useful.}

\item{model,x,y}{logicals. If \code{TRUE} the corresponding components of the fit (the model frame, the model matrix, the response)
 are returned.
\eqn{\beta}{beta}. This argument is ignored for a flat prior.}

\item{center}{logical or numeric. If \code{TRUE} then the covariates will be centered on their means to make them
orthoganal to the intercept. This probably makes no sense for models with factors, and if the argument
is numeric then it contains a vector of covariate indices to be centered (not implemented yet).}
}
\value{
\code{bayes.lm} returns an object of class \code{c("Bolstad.lm", "lm")}.
The functions \code{summary} and \code{link[stats]{anova}} are used to obtain and print a
summary and analysis of variance table of the results. The generic accessor functions \code{coefficients, effects, fitted.values and residuals}
extract various useful features of the value returned by \code{bayes.lm}. Note that the residuals
are computed at the posterior mean values of the coefficients.


An object of class "lm" is a list containing at least the following components:
\item{coefficients	} a named vector of coefficients which contains the posterior mean
\item{post.var} a matrix containing the posterior variance-covariance matrix of the coefficients
\item{post.sd} sigma.
\item{residuals} the residuals, that is response minus fitted values (computed at the posterior mean).
\item{fitted.values} the fitted mean values (computed at the posterior mean).
\item{df.residual} the residual degrees of freedom.
\item{call} the matched call.
\item{terms} the \code{\link[stats]{terms}} object used.
\item{y} if requested, the response used.
\item{x} if requested, the model matrix used.
\item{model} if requested (the default), the model frame used.
\item{na.action} (where relevant) information returned by \code{model.frame} on the special
handling of \code{NA}s.
}
\description{
bayes.lm is used to fit linear models in the Bayesian paradigm. It can be used to carry out regression,
single stratum analysis of variance and analysis of covariance (although these are not tested). This
documentation is shamelessly adapated from the lm documentation
}
\details{
Models for \code{bayes.lm} are specified symbolically. A typical model has the form
\code{response ~ terms} where \code{response} is the (numeric) response vector and \code{terms} is a
 series of terms which specifies a linear predictor for \code{response}. A terms specification of the
 form \code{first + second} indicates all the terms in \code{first} together with all the terms in
 \code{second} with duplicates removed. A specification of the form \code{first:second} indicates the
 set of terms obtained by taking the interactions of all terms in \code{first} with all terms in
 \code{second}. The specification \code{first*second} indicates the cross of \code{first} and \code{second}.
  This is the same as \code{first + second + first:second}.

  See \code{\link[stats]{model.matrix}} for some further details. The terms in the formula will be
  re-ordered so that main effects come first, followed by the interactions, all second-order,
  all third-order and so on: to avoid this pass a \code{terms} object as the formula
  (see \code{\link[stats]{aov}} and \code{demo(glm.vr)} for an example).

  A formula has an implied intercept term. To remove this use either \code{y ~ x - 1} or
  \code{y ~ 0 + x}. See \code{\link[stats]{formula}} for more details of allowed formulae.

  \code{bayes.lm} calls the lower level function \code{lm.fit} to get the maximum likelihood estimates
   see below, for the actual numerical computations. For programming only, you may consider doing
   likewise.

   \code{subset} is evaluated in the same way as variables in formula, that is first in data and
   then in the environment of formula.
}
\examples{
## generate some data from a known model, where the true value of the
## intercept alpha is 2, the true value of the slope beta is 3, and the
## errors come from a normal(0,1) distribution
x = rnorm(50)
y = 22+3*x+rnorm(50)

## use the function with a flat prior for the slope beta and a
## flat prior for the intercept, alpha_xbar.

bayes.lin.reg(y,x)

## use the function with a normal(0,3) prior for the slope beta and a
## normal(30,10) prior for the intercept, alpha_xbar.

bayes.lin.reg(y,x,"n","n",0,3,30,10)

## use the same data but plot it and the credible interval

bayes.lin.reg(y,x,"n","n",0,3,30,10, plot.data = TRUE)

## The heart rate vs. O2 uptake example 14.1
O2 = c(0.47,0.75,0.83,0.98,1.18,1.29,1.40,1.60,1.75,1.90,2.23)
HR = c(94,96,94,95,104,106,108,113,115,121,131)
plot(HR,O2,xlab="Heart Rate",ylab="Oxygen uptake (Percent)")

bayes.lin.reg(y,x,"n","f",0,1,sigma=0.13)

## Repeat the example but obtain predictions for HR = 100 and 110

bayes.lin.reg(y,x,"n","f",0,1,sigma=0.13,pred.x=c(100,110))
}
\keyword{misc}

