\name{Bolstad-package}
\alias{Bolstad-package}
\alias{Bolstad}
\docType{package}
\title{Bolstad Functions
}
\description{
A set of R functions and data sets for the book Introduction to Bayesian Statistics, Bolstad, W.M. (2007), John Wiley & Sons ISBN 0-471-27020-2. Most of the package functions replicate the Minitab macros that are provided with the book. Some additional functions are provided to simplfy inference about the posterior distribution of the parameters of interest.
}
\details{
  \tabular{ll}{
    Package: \tab Bolstad\cr
    Type: \tab Package\cr
    Version: \tab 0.2-26\cr
    Date: \tab 2015-05-01\cr
    License: \tab GPL 2\cr
  }
}
\author{
James Curran
Maintainer: James Curran <j.curran@auckland.ac.nz>
~~ The author and/or maintainer of the package ~~
}
\references{
  Bolstad, W.M. (2007), Introduction to Bayesian Statistics, John Wiley & Sons.
}
\keyword{ package }
