% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bolstad.control.R
\name{Bolstad.control}
\alias{Bolstad.control}
\title{Control Bolstad functions}
\usage{
Bolstad.control(plot = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{plot}{if \code{TRUE} then draw a plot (for functions that actually have plots)}

\item{quiet}{if \code{TRUE} then suppress the function output}

\item{...}{additional parameters}
}
\value{
an invisible list of options and their values
}
\description{
Control Bolstad functions
}
