% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\alias{cdf.Bolstad}
\title{Cumulative distribution function generic}
\usage{
cdf(x, ...)

\method{cdf}{Bolstad}(x, ...)
}
\arguments{
\item{x}{An object for which we want to compute the cdf}

\item{\dots}{Any other parameters. Not currently used.}
}
\value{
either the exact cdf of the posterior if a conjugate prior has been
  used, or a a \code{stats::splinefun} which will compute the lower
  tail probability of the parameter for any valid input.
}
\description{
This function returns the cumulative distribution function (cdf) of the
posterior distribution of the parameter interest over the range of values for
which the posterior is specified.
}
\section{Methods (by class)}{
\itemize{
\item \code{Bolstad}: Cumulative distribution function for posterior density
}}
\author{
James Curran
}

