% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrior.default.R
\name{createPrior.default}
\alias{createPrior.default}
\title{Create prior default method}
\usage{
\method{createPrior}{default}(x, wt, ...)
}
\arguments{
\item{x}{a vector of x values at which the prior is to be specified (the support of the prior).}

\item{wt}{a vector of weights corresponding to the weight of the prior at the
given x values.}

\item{\dots}{optional exta arguments. Not currently used.}
}
\value{
a linear interpolation function where the weights have been scaled so
  the function (numerically) integrates to 1.
}
\description{
Create prior default method
}

