% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createPrior.R
\name{createPrior}
\alias{createPrior}
\alias{createPrior.default}
\title{Create prior generic}
\usage{
createPrior(x, ...)

\method{createPrior}{default}(x, wt, ...)
}
\arguments{
\item{wt}{a vector of weights corresponding to the weight of the prior at the
given x values}
}
\value{
a linear interpolation function where the weights have been scaled so
  the function (numerically) integrates to 1.
}
\description{
Create prior generic
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 
}}

