% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IQR.R
\name{IQR}
\alias{IQR}
\alias{IQR.Bolstad}
\alias{IQR.default}
\title{Interquartile Range generic}
\usage{
IQR(x, ...)

\method{IQR}{default}(x, ...)

\method{IQR}{Bolstad}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{\dots}{any additional arguments. These are primarily used in \code{IQR.default}
which calls \code{stats::IQR}.}
}
\description{
Compute the interquartile range.
}
\details{
If \code{x} is an object of class \code{Bolstad} then the posterior
  median of the parameter of interest will be calculated.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 

\item \code{Bolstad}: 
}}
\author{
James Curran
}

