% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lines.Bolstad.R
\name{lines.Bolstad}
\alias{lines.Bolstad}
\title{Lines method for Bolstad objects}
\usage{
\method{lines.Bolstad}{}(x, ...)
}
\arguments{
\item{x}{an object of class \code{Bolstad}.}

\item{\dots}{any additional parameters to be passed to \code{graphics::lines}.}
}
\description{
Allows simple addition of posterior distributions from other results to an
existing plot
}

