% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/median.Bolstad.R
\name{median.Bolstad}
\alias{median.Bolstad}
\title{Compute the posterior median}
\usage{
\method{median.Bolstad}{}(x, ...)
}
\arguments{
\item{x}{an object of class \code{Bolstad}.}

\item{\dots}{currently ignored}
}
\description{
Compute the posterior median
}
\author{
James Curran
}

