% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moisture.df}
\alias{moisture.df}
\title{Moisture data}
\format{
A data frame with 25 observations on the following 6 variables.
\itemize{ \item{batch. the batch number of the measurement}
\item{proc.level. the in-process moisture level} \item{final.level. natural
the final moisture level of the batch} \item{ls.fit the least squares fitted value
of final.level given proc.level} \item{residiual. the least squares residual}
\item{residiual.sq. the squaredvleast squares residual}}
}
\usage{
moisture.df
}
\description{
Moisture level at two stages in a food manufacturing process, in-process and final.
These data are given in Example 14.1
}
\examples{

data(moisture.df)
plot(final.level~proc.level, data = moisture.df)


}
\keyword{datasets}
