% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sintegral.R
\name{print.sintegral}
\alias{print.sintegral}
\title{Generic print method}
\usage{
\method{print}{sintegral}(x, ...)
}
\arguments{
\item{x}{An object of type \code{sintegral}--see \code{\link{sintegral}}.}

\item{\dots}{other parameters passed to \code{paste0}.}
}
\description{
Print the value of an \code{sintegral} object, specifically the 
(approximate) value of the integral.
}
