% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.Bolstad.R
\name{quantile.Bolstad}
\alias{quantile.Bolstad}
\title{Compute quantiles from the posterior distribution}
\usage{
\method{quantile}{Bolstad}(x, probs = seq(0, 1, 0.25), ...)
}
\arguments{
\item{x}{an object of class \code{Bolstad}}

\item{probs}{numeric vector of probabilities with values in \eqn{[0,1]}.}

\item{\dots, }{any extra arguments needed.}
}
\description{
Compute quantiles from the posterior distribution
}

