% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd}
\alias{sd}
\title{Standard deviation generic}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{an object.}

\item{\dots}{Any additional arguments to be passed to \code{sd}.}
}
\description{
Standard deviation generic
}

