% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPrior.R
\name{createPrior}
\alias{createPrior}
\title{Create prior generic}
\usage{
createPrior(x, ...)
}
\arguments{
\item{x}{a vector of x values at which the prior is to be specified (the support of the prior).}

\item{\dots}{optional exta arguments. Not currently used.}
}
\value{
a linear interpolation function where the weights have been scaled so
  the function (numerically) integrates to 1.
}
\description{
Create prior generic
}
