% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{IQR}
\alias{IQR}
\title{Interquartile Range generic}
\usage{
IQR(x, ...)
}
\arguments{
\item{x}{an object.}

\item{\dots}{any additional arguments. These are primarily used in \code{IQR.default}
which calls \code{stats::IQR}.}
}
\description{
Compute the interquartile range.
}
\details{
If \code{x} is an object of class \code{Bolstad} then the posterior 
  IQR of the parameter of interest will be calculated.
}
\author{
James Curran
}
