% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.Bolstad.R
\name{median.Bolstad}
\alias{median.Bolstad}
\title{Median generic}
\usage{
\method{median}{Bolstad}(x, na.rm = FALSE)
}
\arguments{
\item{x}{an object.}

\item{na.rm}{Ideally if \code{TRUE} then missing values will be removed, but not currently used}
}
\description{
Median generic
}
\details{
If \code{x} is an object of class \code{Bolstad} then the posterior 
  median of the parameter of interest will be calculated.
}
\author{
James Curran
}

