% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.sscamp.R
\name{print.sscsamp}
\alias{print.sscsamp}
\title{Print method for objects of class \code{sscsample}}
\usage{
\method{print}{sscsamp}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sscsamp} produced by \code{sscsample}}

\item{\dots}{any other arguments that are to be passed to \code{cat}}
}
\description{
This function provides a print summary method for the output of
\code{sscsample}. The \code{sscsample} produces a large number of samples
from a fixed population using either simple random, stratified, or cluster
sampling. This function provides the means of each sample plus the number of
observations from each ethnicity stratum in the sample.
}
\author{
James Curran
}
\seealso{
\code{\link{sscsample}}
}

