% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sd.R
\name{sd}
\alias{sd}
\alias{sd.Bolstad}
\alias{sd.default}
\title{Standard deviation generic}
\usage{
sd(x, ...)

\method{sd}{default}(x, ...)

\method{sd}{Bolstad}(x, ...)
}
\arguments{
\item{x}{An object for which we want to compute the standard deviation.}

\item{\dots}{Any additional arguments to be passed to \code{sd}.}
}
\description{
Standard deviation generic
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Makes sure that the base \code{stats::sd} gets called for anything
other than an object of class \code{Bolstad}.

\item \code{Bolstad}: Calculate the posterior mean of an object of class \code{Bolstad}. If the
 object has a member \code{sd} then it will return this value otherwise it
 will calculate the posterior standard deviation \eqn{sd[\theta|x]} using
 linear interpolation to approximate the density function and numerical
 integration where \eqn{\theta} is the variable for which we want to do
 Bayesian inference, and \eqn{x} is the data.
}}
\examples{
# The useful of this method is really highlighted when we have a general
# continuous prior. In this example we are interested in the posterior
# standard deviation of an normal mean. Our prior is triangular over [-3, 3]
set.seed(123)
x = rnorm(20, -0.5, 1)

mu = seq(-3, 3, by = 0.001)

mu.prior = rep(0, length(mu))
mu.prior[mu <= 0] = 1 / 3 + mu[mu <= 0] / 9
mu.prior[mu > 0] = 1 / 3 - mu[mu > 0] / 9

results = normgcp(x, 1, density = "user", mu = mu, mu.prior = mu.prior, plot = FALSE)
sd(results)
}

